+++ packages/camera/camera_platform_interface/lib/src/utils/utils.dart
@@ -37,13 +37,13 @@ String serializeDeviceOrientation(DeviceOrientation orientation) {
 /// Returns the device orientation for a given String.
 DeviceOrientation deserializeDeviceOrientation(String str) {
   switch (str) {
-    case "portraitUp":
+    case 'portraitUp':
       return DeviceOrientation.portraitUp;
-    case "portraitDown":
+    case 'portraitDown':
       return DeviceOrientation.portraitDown;
-    case "landscapeRight":
+    case 'landscapeRight':
       return DeviceOrientation.landscapeRight;
-    case "landscapeLeft":
+    case 'landscapeLeft':
       return DeviceOrientation.landscapeLeft;
     default:
       throw ArgumentError('"$str" is not a valid DeviceOrientation value');