+++ packages/video_player/video_player_web/CHANGELOG.md
@@ -1,7 +1,12 @@
+## 2.0.5
+
+* Adds compatibility with `video_player_platform_interface` 5.0, which does not
+  include non-dev test dependencies.
+
 ## 2.0.4
 
 * Adopt `video_player_platform_interface` 4.2 and opt out of `contentUri` data source.
-  
+
 ## 2.0.3
 
 * Add `implements` to pubspec.